+++ src/Library/Clases/UserAdmin.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 
 namespace Proyecto_Final
 {
@@ -8,6 +9,10 @@ namespace Proyecto_Final
     public class UserAdmin
     {
 
+        /// <summary>
+        /// Otorga el id del administrador.
+        /// </summary>
+        /// <value></value>
         public string Id { get; set; }
 
         /// <summary>
@@ -19,6 +24,7 @@ namespace Proyecto_Final
         /// <summary>
         /// Inicializa la clase UserAdmin.
         /// </summary>
+        /// <param name="id"></param>
         /// <param name="nombre"></param>
         public UserAdmin(string id, string nombre)
         {
@@ -27,18 +33,24 @@ namespace Proyecto_Final
         }
 
         /// <summary>
-        /// Invita a una empresa desde cualquier IUserInterface siempre y cuando esta empresa no haya sido invitada.
+        /// Genera un token de invitacion para ser enviado y lo almacena para su verificacion.
         /// </summary>
-        /// <param name="nombreUserEmpresa"></param>
-        public void InvitarEmpresa(string nombreUserEmpresa)
+        /// <returns>Devuelve un token generado como string</returns>
+        public string InvitarEmpresa()
         {
-            foreach (UserEmpresa user in Singleton<Datos>.Instance.ListaUsuarioEmpresa())
-            {
-                if (user.Nombre == nombreUserEmpresa)
-                {
-                    user.Invitacion = new Invitacion();
-                }
-            }
+            return this.generateToken();
+        }
+
+        private string generateToken()
+        {
+            string allChar = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";  
+            Random random = new Random();  
+            string resultToken = new string(  
+            Enumerable.Repeat(allChar , 16)  
+                        .Select(token => token[random.Next(token.Length)]).ToArray());   
+   
+            string authToken = resultToken.ToString();  
+            return authToken;
         }
     }
 }
