+++ src/Library/Handlers/InviteHandler.cs
@@ -48,12 +48,37 @@ namespace Proyecto_Final
                 {
                     if (Singleton<Datos>.Instance.IsAdmin(message.UserId))
                     {
-                        string newToken = this.generateToken();
-                        response = $"Envie este token a la persona que desea invitar como Empresa.\n\nTOKEN:\n{newToken}";
-                        Singleton<Datos>.Instance.AgregarToken(newToken);
-                        Console.WriteLine($"InviteHandler: TOKEN | {newToken} | generado por < {message.UserId} > ");
-                        Singleton<StatusManager>.Instance.AgregarEstadoUsuario(message.UserId, "STATUS_IDLE");
-                        return true;
+                        if (Singleton<Datos>.Instance.IsRegistered(message.UserId))
+                        {
+                            foreach (UserAdmin admin in Singleton<Datos>.Instance.ListaUsuariosRegistrados())
+                            {
+                                if (admin.Id == message.UserId)
+                                {
+                                    string newToken = admin.InvitarEmpresa();
+                                    response = $"Envie este token a la persona que desea invitar como Empresa.\n\nTOKEN:\n{newToken}";
+                                    Singleton<Datos>.Instance.AgregarToken(newToken);
+                                    Console.WriteLine($"InviteHandler: TOKEN | {newToken} | generado por < {message.UserId} > ");
+                                    Singleton<StatusManager>.Instance.AgregarEstadoUsuario(message.UserId, "STATUS_IDLE");
+                                    return true;
+                                }
+                            }
+                        }
+                        else
+                        {
+                            Singleton<UserCreator>.Instance.CrearUserAdmin(message.UserId, message.Text); // Tiene que ir el registro de admin aqui?
+                            foreach (UserAdmin admin in Singleton<Datos>.Instance.ListaUsuariosRegistrados())
+                            {
+                                if (admin.Id == message.UserId)
+                                {
+                                    string newToken = admin.InvitarEmpresa();
+                                    response = $"Envie este token a la persona que desea invitar como Empresa.\n\nTOKEN:\n{newToken}";
+                                    Singleton<Datos>.Instance.AgregarToken(newToken);
+                                    Console.WriteLine($"InviteHandler: TOKEN | {newToken} | generado por < {message.UserId} > ");
+                                    Singleton<StatusManager>.Instance.AgregarEstadoUsuario(message.UserId, "STATUS_IDLE");
+                                    return true;
+                                }
+                            }
+                        }
                     }
                     else
                     {
@@ -65,17 +90,6 @@ namespace Proyecto_Final
             }
             response = string.Empty;
             return false;
-        }
-        private string generateToken()
-        {
-            string allChar = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";  
-            Random random = new Random();  
-            string resultToken = new string(  
-            Enumerable.Repeat(allChar , 16)  
-                        .Select(token => token[random.Next(token.Length)]).ToArray());   
-   
-            string authToken = resultToken.ToString();  
-            return authToken;
         }
     }
